python/advent/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "advent"
version = "0.0.1"
dependencies = [
    "python-dotenv >= 1.0.0",
]
requires-python = ">=3.12"

[project.scripts]
advent = "advent.__main__:main"
